sw: remove SetAutoFormatLvl calls in SwAutoFormat
BuildTextIndent() calls SetAutoFormatLvl() if bAFormatByInput,
while also setting the node's style to RES_POOLCOLL_TEXT_MOVE.

Then CalcLevel short-cuts if the node has the style
RES_POOLCOLL_TEXT_MOVE and bAFormatByInput is set, and it also
resets via SetAutoFormatLvl(0).

So BuildTextIndent() affects the next caller of CalcLevel() on that
node, but not the 2nd or later callers.

Of the callers of BuildTextIndent(), one of them will actually call
CalcLevel() immediately, while the others all set the state to
READ_NEXT_PARA; since the iteration is always forward, CalcLevel() will
not be called on that node while the SwAutoFormat instance is alive.

It doesn't make sense to affect the next SwAutoFormat invocation but not
the one after that?  So just replace this with a return value from
BuildTextIndent() that is checked in the one immediate use location.

Change-Id: I548358c508ac65cc07d13da6708e285c910e1663