sw_redlinehide_4a: SwAutoFormat iterates frames, not nodes
In order to not require moving the delete redlines out of the document
body, rely on the DeleteSel() implementation skipping over the redlines,
and also pass the shell's layout to the various SwDoc functions that
are called.

Previously, the redlines were not deleted by DeleteSel() due to
SwAutoFormat setting RedlineFlags::Ignore.

Change-Id: Ib05ecb534993368d81f66e70124617968b4c8e9a